Add global utility module with reused functions.
